+++ src/main/java/blackjack/domain/card/HoldCards.java
@@ -1,4 +1,7 @@
-package blackjack;
+package blackjack.domain.card;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Number;
 
 import java.util.ArrayList;
 import java.util.List;
