+++ src/main/java/blackjack/domain/entry/Participant.java
@@ -1,4 +1,7 @@
-package blackjack;
+package blackjack.domain.entry;
+
+import blackjack.domain.card.HoldCards;
+import blackjack.domain.card.Card;
 
 import java.util.List;
