+++ src/test/java/blackjack/CommandTest.java
@@ -1,5 +1,6 @@
 package blackjack;
 
+import blackjack.domain.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
