+++ src/test/java/blackjack/DeckTest.java
@@ -1,5 +1,7 @@
 package blackjack;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
