+++ src/test/java/blackjack/OutcomeTest.java
@@ -1,5 +1,6 @@
 package blackjack;
 
+import blackjack.domain.Outcome;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
